Updated void initialize_card_on_table(void) - to ensure the first card on table is normal card.
